requirements.txt
playwright
requests
selectolax
Pillow
schedule
g4f